refactor: package.json type:module — drop .mts cascade, simplify throttle
Root cause fix for the .mts cascade introduced in #4 and grown in
this slice: the underlying need was top-level await in tsx scripts,
which tsx blocks because it infers CJS for .ts when package.json has
no "type" field. Setting "type": "module" makes tsx treat .ts as
ESM, so top-level await works without the .mts workaround.

With that one line in package.json:

- All scripts/crawl/*.mts → .ts (6 files), plus scripts/blob/upload-
  fixture.mts → .ts. Test imports drop the .mjs extension workaround
  required by TS bundler resolution for .ts → .mts. `<T,>` trailing
  commas drop back to plain `<T>` (only .mts/.cts needed the comma).
- Crawler files now match the rest of the codebase (chart-schema.ts,
  charts-client.ts, etc.) on extension and import shape.

Throttle module: rpm parameter had no defensible reason — only call
site in prod is createThrottle(20), tests don't exercise other rpms
meaningfully. Replaced with a module-level GAP_MS = 3000 constant
with a comment pointing at Apple's 20/min/IP source.

Throttle tests: tightened the timing assertions. The previous
"advance 3000ms, expect resolved" pattern passed even if the actual
gap was 100ms. Switched to a resolved-flag pattern that asserts the
exact boundary: pending after advancing 2999ms, resolved after the
3000th. Same five behavioural cases, stricter signal.

Verified end-to-end: pnpm crawl kr re-runs cleanly against live
Apple APIs and Vercel Blob; 33 tests + lint + typecheck pass.
