+++ scripts/crawl/itunes-lookup.test.ts
@@ -2,7 +2,7 @@ import { readFile } from "node:fs/promises";
 import { dirname, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
-import { ItunesLookupError, lookupTrack } from "./itunes-lookup.mjs";
+import { ItunesLookupError, lookupTrack } from "./itunes-lookup";
 
 const FIXTURE_PATH = resolve(
   dirname(fileURLToPath(import.meta.url)),
