+++ scripts/crawl/revalidate-trigger.ts
